moved the machine hooks to hooks/
