+++ src/components/community/Platforms/components/JoinModal.tsx
@@ -15,7 +15,7 @@ import { ArrowSquareOut } from "phosphor-react"
 import QRCode from "qrcode.react"
 import { Error } from "components/common/Error"
 import { useCommunity } from "components/community/Context"
-import useJoinModalMachine from "../utils/joinModalMachineConfig"
+import useJoinModalMachine from "../hooks/useJoinModalMachine"
 import { usePersonalSign } from "../hooks/usePersonalSign"
 import platformsContent from "../platformsContent"
 import processSignError from "../utils/processSignError"
