+++ src/components/community/Platforms/hooks/useLeaveModalMachine.ts
@@ -1,5 +1,5 @@
 import useStateMachine from "@cassiozen/usestatemachine"
-import type { SignErrorType } from "../hooks/usePersonalSign"
+import type { SignErrorType } from "./usePersonalSign"
 
 type MachineContextType = {
   error: SignErrorType | null
